requirements.txt
--extra-index-url https://download.pytorch.org/whl/lts/1.8/cu111
supervisely==6.69.78
torch==1.8.2
torchvision==0.9.2
Pillow==8.4.0
Cython==0.29.23
mmcv-full==1.4.1
gdown
attr